requirements.txt
openai>=1.0.0
langchain>=0.1.0
python-oracledb>=2.0.0
python-dotenv>=1.0.0